Update unit tests + code review updates
