+++ plugins/module_utils/ec2.py
@@ -43,11 +43,9 @@ from typing import Any
 from typing import Dict
 from typing import List
 from typing import Optional
-from typing import Tuple
 from typing import Union
 
 import ansible.module_utils.common.warnings as ansible_warnings
-from ansible.module_utils._text import to_native
 from ansible.module_utils.ansible_release import __version__
 
 # Used to live here, moved into ansible.module_utils.common.dict_transformations
@@ -77,7 +75,6 @@ from .exceptions import AnsibleAWSError  # pylint: disable=unused-import
 
 # Used to live here, moved into ansible_collections.amazon.aws.plugins.module_utils.modules
 # The names have been changed in .modules to better reflect their applicability.
-from .modules import AnsibleAWSModule
 from .modules import _aws_common_argument_spec as aws_common_argument_spec  # pylint: disable=unused-import
 from .modules import aws_argument_spec as ec2_argument_spec  # pylint: disable=unused-import
 
@@ -162,7 +159,7 @@ def disassociate_subnet_cidr_block(client, association_id: str) -> None:
 @EC2VpcSubnetErrorHandler.common_error_handler("associate subnet cidr block")
 @AWSRetry.jittered_backoff()
 def associate_subnet_cidr_block(client, subnet_id: str, **params) -> Dict[str, Any]:
-    client.associate_subnet_cidr_block(SubnetId=subnet_id, **params)["Ipv6CidrBlockAssociation"]
+    return client.associate_subnet_cidr_block(SubnetId=subnet_id, **params)["Ipv6CidrBlockAssociation"]
 
 
 # EC2 VPC Route table
@@ -305,11 +302,9 @@ class EC2InternetGatewayErrorHandler(AWSErrorHandler):
         return is_boto3_error_code("InvalidInternetGatewayID.NotFound")
 
 
-@EC2InternetGatewayErrorHandler.list_error_handler("describe internet gateways", [])
 @AWSRetry.jittered_backoff()
 def describe_internet_gateways(client, **params: Dict[str, Any]) -> List[Dict[str, Any]]:
-    paginator = client.get_paginator("describe_internet_gateways")
-    return paginator.paginate(**params).build_full_result()["InternetGateways"]
+    return client.describe_internet_gateways(**params)["InternetGateways"]
 
 
 @EC2InternetGatewayErrorHandler.common_error_handler("create internet gateway")
@@ -568,7 +563,7 @@ class EC2InstanceErrorHandler(AWSErrorHandler):
 
 
 @EC2InstanceErrorHandler.list_error_handler("describe instances", [])
-@AWSRetry.jittered_backoff(catch_extra_error_codes=EC2_INSTANCE_CATCH_EXTRA_CODES)
+@AWSRetry.jittered_backoff()
 def describe_instances(client, **params) -> List[Dict[str, Any]]:
     paginator = client.get_paginator("describe_instances")
     return paginator.paginate(**params).build_full_result()["Reservations"]
@@ -1202,421 +1197,3 @@ def normalize_ec2_vpc_dhcp_config(option_config):
                 config_data[option] = [val["Value"] for val in config_item["Values"]]
 
     return config_data
-
-
-def get_default_vpc(client, module: AnsibleAWSModule) -> Optional[Dict[str, Any]]:
-    try:
-        vpcs = describe_vpcs(client, Filters=ansible_dict_to_boto3_filter_list({"isDefault": "true"}))
-    except AnsibleEC2Error as e:
-        module.fail_json_aws(e, msg="Could not describe default VPC")
-    return None if len(vpcs) == 0 else vpcs[0]
-
-
-def get_default_subnet(
-    client, module: AnsibleAWSModule, vpc: Dict[str, Any], availability_zone: Optional[str] = None
-) -> Optional[Dict[str, Any]]:
-    try:
-        subnets = describe_subnets(
-            client,
-            Filters=ansible_dict_to_boto3_filter_list(
-                {
-                    "vpc-id": vpc["VpcId"],
-                    "state": "available",
-                    "default-for-az": "true",
-                }
-            ),
-        )
-    except AnsibleEC2Error as e:
-        module.fail_json_aws(e, msg=f"Could not describe default subnets for VPC {vpc['VpcId']}")
-    if subnets:
-        if availability_zone is not None:
-            match = [subnet for subnet in subnets if subnet["AvailabilityZone"] == availability_zone]
-            if match:
-                return match[0]
-
-        # to have a deterministic sorting order, we sort by AZ so we'll always pick the `a` subnet first
-        # there can only be one default-for-az subnet per AZ, so the AZ key is always unique in this list
-        by_az = sorted(subnets, key=lambda s: s["AvailabilityZone"])
-        return by_az[0]
-    return None
-
-
-# EC2 instances modules utilities
-class EC2InstanceModule(AnsibleAWSModule):
-    _ANSIBLE_STATE_TO_EC2_MAPPING = {
-        "present": {"waiter": "instance_exists", "state": "running"},
-        "started": {"waiter": "instance_status_ok", "state": "running"},
-        "running": {"waiter": "instance_running", "state": "running"},
-        "stopped": {"waiter": "instance_stopped", "state": "stopped"},
-        "restarted": {"waiter": "instance_status_ok", "state": "running"},
-        "rebooted": {"waiter": "instance_running", "state": "running"},
-        "terminated": {"waiter": "instance_terminated", "state": "terminated"},
-        "absent": {"waiter": "instance_terminated", "state": "terminated"},
-    }
-
-    def __init__(self) -> None:
-        super().__init__(
-            argument_spec=self.argument_specs(),
-            mutually_exclusive=[
-                ["security_groups", "security_group"],
-                ["availability_zone", "vpc_subnet_id"],
-                ["aap_callback", "user_data"],
-                ["image_id", "image"],
-                ["exact_count", "count"],
-                ["exact_count", "instance_ids"],
-                ["tenancy", "placement"],
-                ["placement_group", "placement"],
-            ],
-            supports_check_mode=True,
-        )
-
-        if self.params.get("network"):
-            if self.params.get("network").get("interfaces"):
-                if self.params.get("security_group"):
-                    self.fail_json(msg="Parameter network.interfaces can't be used with security_group")
-                if self.params.get("security_groups"):
-                    self.fail_json(msg="Parameter network.interfaces can't be used with security_groups")
-
-        if self.params.get("placement_group"):
-            self.deprecate(
-                "The placement_group parameter has been deprecated, please use placement.group_name instead.",
-                date="2025-12-01",
-                collection_name="amazon.aws",
-            )
-
-        if self.params.get("tenancy"):
-            self.deprecate(
-                "The tenancy parameter has been deprecated, please use placement.tenancy instead.",
-                date="2025-12-01",
-                collection_name="amazon.aws",
-            )
-
-        self.ec2_client = self.client("ec2")
-
-    @property
-    def ec2(self):
-        return self.ec2_client
-
-    @staticmethod
-    def argument_specs() -> Dict[str, Any]:
-        # running/present are synonyms
-        # as are terminated/absent
-        return dict(
-            state=dict(
-                default="present",
-                choices=["present", "started", "running", "stopped", "restarted", "rebooted", "terminated", "absent"],
-            ),
-            wait=dict(default=True, type="bool"),
-            wait_timeout=dict(default=600, type="int"),
-            count=dict(type="int"),
-            exact_count=dict(type="int"),
-            image=dict(type="dict"),
-            image_id=dict(type="str"),
-            instance_type=dict(type="str"),
-            user_data=dict(type="str"),
-            aap_callback=dict(
-                type="dict",
-                aliases=["tower_callback"],
-                required_if=[
-                    (
-                        "windows",
-                        False,
-                        (
-                            "tower_address",
-                            "job_template_id",
-                            "host_config_key",
-                        ),
-                        False,
-                    ),
-                ],
-                options=dict(
-                    windows=dict(type="bool", default=False),
-                    set_password=dict(type="str", no_log=True),
-                    tower_address=dict(type="str"),
-                    job_template_id=dict(type="str"),
-                    host_config_key=dict(type="str", no_log=True),
-                ),
-            ),
-            ebs_optimized=dict(type="bool"),
-            vpc_subnet_id=dict(type="str", aliases=["subnet_id"]),
-            availability_zone=dict(type="str"),
-            security_groups=dict(default=[], type="list", elements="str"),
-            security_group=dict(type="str"),
-            iam_instance_profile=dict(type="str", aliases=["instance_role"]),
-            name=dict(type="str"),
-            tags=dict(type="dict", aliases=["resource_tags"]),
-            purge_tags=dict(type="bool", default=True),
-            filters=dict(type="dict", default=None),
-            launch_template=dict(type="dict"),
-            license_specifications=dict(
-                type="list",
-                elements="dict",
-                options=dict(
-                    license_configuration_arn=dict(type="str", required=True),
-                ),
-            ),
-            key_name=dict(type="str"),
-            cpu_credit_specification=dict(type="str", choices=["standard", "unlimited"]),
-            cpu_options=dict(
-                type="dict",
-                options=dict(
-                    core_count=dict(type="int", required=True),
-                    threads_per_core=dict(type="int", choices=[1, 2], required=True),
-                ),
-            ),
-            tenancy=dict(type="str", choices=["dedicated", "default"]),
-            placement_group=dict(type="str"),
-            placement=dict(
-                type="dict",
-                options=dict(
-                    affinity=dict(type="str"),
-                    availability_zone=dict(type="str"),
-                    group_name=dict(type="str"),
-                    host_id=dict(type="str"),
-                    host_resource_group_arn=dict(type="str"),
-                    partition_number=dict(type="int"),
-                    tenancy=dict(type="str", choices=["dedicated", "default", "host"]),
-                ),
-            ),
-            instance_initiated_shutdown_behavior=dict(type="str", choices=["stop", "terminate"]),
-            termination_protection=dict(type="bool"),
-            hibernation_options=dict(type="bool", default=False),
-            detailed_monitoring=dict(type="bool"),
-            instance_ids=dict(default=[], type="list", elements="str"),
-            network=dict(default=None, type="dict"),
-            volumes=dict(default=None, type="list", elements="dict"),
-            metadata_options=dict(
-                type="dict",
-                options=dict(
-                    http_endpoint=dict(choices=["enabled", "disabled"], default="enabled"),
-                    http_put_response_hop_limit=dict(type="int", default=1),
-                    http_tokens=dict(choices=["optional", "required"], default="optional"),
-                    http_protocol_ipv6=dict(choices=["disabled", "enabled"], default="disabled"),
-                    instance_metadata_tags=dict(choices=["disabled", "enabled"], default="disabled"),
-                ),
-            ),
-            additional_info=dict(type="str"),
-        )
-
-    def get_default_vpc(self) -> Optional[Dict[str, Any]]:
-        return get_default_vpc(self.ec2, self)
-
-    def get_default_subnet(
-        self, vpc: Dict[str, Any], availability_zone: Optional[str] = None
-    ) -> Optional[Dict[str, Any]]:
-        return get_default_subnet(client=self.ec2, module=self, vpc=vpc, availability_zone=availability_zone)
-
-    def build_filters(self) -> Dict[str, Any]:
-        filters = {
-            # all states except shutting-down and terminated
-            "instance-state-name": ["pending", "running", "stopping", "stopped"],
-        }
-        instance_ids = self.params.get("instance_ids")
-        if isinstance(instance_ids, string_types):
-            filters["instance-id"] = [instance_ids]
-        elif isinstance(instance_ids, list) and len(instance_ids) > 0:
-            filters["instance-id"] = instance_ids
-        else:
-            vpc_subnet_id = self.params.get("vpc_subnet_id")
-            filters["subnet-id"] = [vpc_subnet_id]
-            if not vpc_subnet_id:
-                network = self.params.get("network")
-                if network:
-                    # grab AZ from one of the ENIs
-                    interfaces = network.get("interfaces")
-                    if interfaces:
-                        filters["network-interface.network-interface-id"] = []
-                        for i in interfaces:
-                            if isinstance(i, dict):
-                                i = i["id"]
-                            filters["network-interface.network-interface-id"].append(i)
-                else:
-                    sub = (
-                        self.get_default_subnet(
-                            self.get_default_vpc() or {},
-                            availability_zone=self.params.get("availability_zone"),
-                        )
-                        or {}
-                    )
-                    filters["subnet-id"] = sub.get("SubnetId")
-
-            name = self.params.get("name")
-            tags = self.params.get("tags")
-            if name:
-                filters["tag:Name"] = [name]
-            elif tags:
-                name_tag = tags.get("Name")
-                if name_tag:
-                    filters["tag:Name"] = [name_tag]
-
-            image_id = self.params.get("image_id")
-            image = self.params.get("image")
-            if image_id:
-                filters["image-id"] = [image_id]
-            elif image and image.get("id"):
-                filters["image-id"] = [image.get("id")]
-        return filters
-
-    def find_instances(
-        self, ids: Optional[List[str]] = None, filters: Optional[Dict[str, Any]] = None
-    ) -> List[Dict[str, Any]]:
-        sanitized_filters = dict()
-        if ids:
-            params = {"InstanceIds": ids}
-        elif filters is None:
-            self.fail_json(msg="No filters provided when they were required")
-        else:
-            for key in list((filters or {}).keys()):
-                if not key.startswith("tag:"):
-                    sanitized_filters[key.replace("_", "-")] = (filters or {}).get(key)
-                else:
-                    sanitized_filters[key] = (filters or {}).get(key)
-            params = {"Filters": ansible_dict_to_boto3_filter_list(sanitized_filters)}
-
-        try:
-            results = []
-            for reservation in describe_instances(self.ec2, **params):
-                results += reservation.get("Instances", [])
-        except AnsibleEC2Error as e:
-            self.fail_json_aws(e, msg="Could not describe instances")
-        return results
-
-    def await_instances(self, ids: List[str], desired_module_state: str = "present", force_wait: bool = False) -> None:
-        if not self.check_mode and ids and (self.params.get("wait", True) or force_wait):
-            boto3_waiter_type = self._ANSIBLE_STATE_TO_EC2_MAPPING.get(desired_module_state, {}).get("waiter")
-            if not boto3_waiter_type:
-                self.fail_json(msg=f"Cannot wait for state {desired_module_state}, invalid state")
-            waiter = self.ec2.get_waiter(boto3_waiter_type)
-            try:
-                waiter.wait(
-                    InstanceIds=ids,
-                    WaiterConfig={
-                        "Delay": 15,
-                        "MaxAttempts": self.params.get("wait_timeout", 600) // 15,
-                    },
-                )
-            except botocore.exceptions.WaiterConfigError as e:
-                instance_ids = ", ".join(ids)
-                self.fail_json(
-                    msg=f"{to_native(e)}. Error waiting for instances {instance_ids} to reach state {boto3_waiter_type}"
-                )
-            except botocore.exceptions.WaiterError as e:
-                instance_ids = ", ".join(ids)
-                self.warn(f"Instances {instance_ids} took too long to reach state {boto3_waiter_type}. {to_native(e)}")
-
-    def _update_instances(self, operation: str, instance_ids: List[str]) -> Tuple[List[str], str]:
-        _OPERATION_MAPPING = {
-            "terminate": "TerminatingInstances",
-            "stop": "StoppingInstances",
-            "start": "StartingInstances",
-        }
-
-        failure_reason = ""
-        result_instance_ids = []
-        try:
-            func = getattr(self.ec2, operation + "_instances")
-            result = func(InstanceIds=instance_ids)
-            result_instance_ids = [i["InstanceId"] for i in result[_OPERATION_MAPPING.get(operation)]]
-        except (botocore.exceptions.ClientError, botocore.exceptions.BotoCoreError) as e:
-            try:
-                failure_reason = to_native(e.message)
-            except AttributeError:
-                failure_reason = to_native(e)
-        return result_instance_ids, failure_reason
-
-    def _terminate_ec2_instances(self, instances: List[Dict[str, Any]]) -> Tuple[List[str], List[str], str]:
-        # Before terminating an instance we need for them to leave
-        # 'pending' or 'stopping' (if they're in those states)
-        wait_status_stopped = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] == "stopping"
-        ]
-        wait_status_running = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] == "pending"
-        ]
-
-        self.await_instances(ids=wait_status_stopped, desired_module_state="stopped", force_wait=True)
-        self.await_instances(ids=wait_status_running, desired_module_state="running", force_wait=True)
-
-        # now terminate instances as they have reached the expected status
-        terminated_instance_ids = [instance["InstanceId"] for instance in instances]
-        failure_reason = ""
-        if not self.check_mode and terminated_instance_ids:
-            terminated_instance_ids, failure_reason = self._update_instances(
-                operation="terminate", instance_ids=terminated_instance_ids
-            )
-        return terminated_instance_ids, [], failure_reason
-
-    def _stop_ec2_instances(self, instances: List[Dict[str, Any]]) -> Tuple[List[str], List[str], str]:
-        # Before stopping an instance we need for them to leave 'pending'
-        wait_status_running = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] == "pending"
-        ]
-        self.await_instances(ids=wait_status_running, desired_module_state="running", force_wait=True)
-
-        # Instances already moving to the relevant state
-        already_stopped_instances = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] in ("stopping", "stopped")
-        ]
-
-        # now stop instances as they have reached the expected status
-        stopped_instance_ids = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] not in ("stopping", "stopped")
-        ]
-        failure_reason = ""
-        if not self.check_mode and stopped_instance_ids:
-            stopped_instance_ids, failure_reason = self._update_instances(
-                operation="stop", instance_ids=stopped_instance_ids
-            )
-
-        return stopped_instance_ids, already_stopped_instances, failure_reason
-
-    def _start_ec2_instances(self, instances: List[Dict[str, Any]]) -> Tuple[List[str], List[str], str]:
-        # Before stopping an instance we need for them to leave 'pending'
-        wait_status_stopped = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] == "stopping"
-        ]
-        self.await_instances(ids=wait_status_stopped, desired_module_state="stopped", force_wait=True)
-
-        # Instances already moving to the relevant state
-        already_running_instances = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] in ("pending", "running")
-        ]
-
-        # Now start instances as they have reached the expected status
-        started_instance_ids = [
-            instance["InstanceId"] for instance in instances if instance["State"]["Name"] not in ("pending", "running")
-        ]
-        failure_reason = ""
-        if not self.check_mode and started_instance_ids:
-            started_instance_ids, failure_reason = self._update_instances(
-                operation="start", instance_ids=started_instance_ids
-            )
-        return started_instance_ids, already_running_instances, failure_reason
-
-    def change_instance_state(
-        self, filters: Optional[List[Dict[str, Any]]], desired_module_state: str
-    ) -> Tuple[List[str], List[str], List[Dict[str, Any]], str]:
-        desired_ec2_state = self._ANSIBLE_STATE_TO_EC2_MAPPING[desired_module_state].get("state")
-        instances = self.find_instances(filters=filters)
-        changed = []
-        change_failed = []
-        failure_reason = ""
-        if instances:
-            to_change = set(
-                instance["InstanceId"] for instance in instances if instance["State"]["Name"] != desired_ec2_state
-            )
-
-            if desired_ec2_state == "terminated":
-                changed, unchanged, failure_reason = self._terminate_ec2_instances(instances)
-            elif desired_ec2_state == "stopped":
-                changed, unchanged, failure_reason = self._stop_ec2_instances(instances)
-            elif desired_ec2_state == "running":
-                changed, unchanged, failure_reason = self._start_ec2_instances(instances)
-
-            if changed:
-                self.await_instances(ids=changed + unchanged, desired_module_state=desired_module_state)
-
-            change_failed = list(to_change - set(changed))
-            changed = list(set(changed))
-            instances = self.find_instances(ids=[i["InstanceId"] for i in instances])
-        return changed, change_failed, instances, failure_reason
