+++ tests/unit/plugins/modules/test_ec2_import_image.py
@@ -1,7 +1,6 @@
 # This file is part of Ansible
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from unittest.mock import ANY
 from unittest.mock import MagicMock
 from unittest.mock import patch
 
@@ -33,46 +32,21 @@ expected_result = {
 
 describe_import_image_tasks = [
     {
-        "ImportTaskId": "import-ami-0c207d759080a3dff",
-        "Progress": "19",
-        "SnapshotDetails": [
+        "import_task_id": "import-ami-0c207d759080a3dff",
+        "progress": "19",
+        "snapshot_details": [
             {
-                "DiskImageSize": 26843545600.0,
-                "Format": "RAW",
-                "Status": "active",
-                "UserBucket": {"S3Bucket": "clone-vm-s3-bucket", "S3Key": "clone-vm-s3-bucket/ubuntu-vm-clone.raw"},
+                "disk_image_size": 26843545600.0,
+                "format": "RAW",
+                "status": "active",
+                "user_bucket": {"s3_bucket": "clone-vm-s3-bucket", "s3_key": "clone-vm-s3-bucket/ubuntu-vm-clone.raw"},
             }
         ],
-        "Status": "active",
-        "StatusMessage": "converting",
-        "Tags": [{"Key": "Name", "Value": "clone-vm-import-image"}],
+        "status": "active",
+        "status_message": "converting",
+        "tags": {"Name": "clone-vm-import-image"},
     }
 ]
-
-
-@pytest.fixture
-def paginate():
-    # Create a MagicMock for the paginate object
-    paginate_mock = MagicMock()
-
-    return paginate_mock
-
-
-@pytest.fixture
-def conn_paginator(paginate):
-    conn_paginator_mock = MagicMock()
-    conn_paginator_mock.paginate.return_value = paginate
-    return conn_paginator_mock
-
-
-@pytest.fixture
-def client(conn_paginator):
-    client_mock = MagicMock()
-
-    # Configure the client.get_paginator to return the conn_paginator
-    client_mock.get_paginator.return_value = conn_paginator
-
-    return client_mock
 
 
 @pytest.fixture
@@ -94,14 +68,14 @@ def module():
 
 
 @pytest.mark.parametrize(
-    "side_effects, expected_result",
+    "import_image_info, expected_result",
     [
         (
-            [{"ImportImageTasks": []}, {"ImportImageTasks": describe_import_image_tasks}],
+            [[], describe_import_image_tasks],
             {"changed": True, "import_image": expected_result},
         ),
         (
-            [{"ImportImageTasks": describe_import_image_tasks}, {"ImportImageTasks": describe_import_image_tasks}],
+            [describe_import_image_tasks, describe_import_image_tasks],
             {
                 "changed": False,
                 "msg": "An import task with the specified name already exists",
@@ -110,26 +84,36 @@ def module():
         ),
     ],
 )
-def test_present_no_check_mode(client, module, paginate, side_effects, expected_result):
-    paginate.build_full_result.side_effect = side_effects
+@patch(module_name + ".import_image")
+@patch(module_name + ".describe_import_image_tasks_as_snake_dict")
+def test_present_no_check_mode(
+    m_describe_import_image_tasks_as_snake_dict, m_import_image, module, import_image_info, expected_result
+):
+    m_describe_import_image_tasks_as_snake_dict.side_effect = import_image_info
     module.exit_json.side_effect = SystemExit(1)
+    connection = MagicMock()
 
-    with patch(utils + ".helper_describe_import_image_tasks", return_value=paginate):
-        with pytest.raises(SystemExit):
-            ec2_import_image.present(client, module)
+    with pytest.raises(SystemExit):
+        ec2_import_image.present(connection, module)
 
+    if import_image_info:
+        m_import_image.call_count = 1
+        m_describe_import_image_tasks_as_snake_dict.call_count = 2
+    else:
+        m_import_image.assert_not_called()
+        m_describe_import_image_tasks_as_snake_dict.call_count = 1
     module.exit_json.assert_called_with(**expected_result)
 
 
 @pytest.mark.parametrize(
-    "side_effects, expected_result",
+    "import_image_info, expected_result",
     [
         (
-            [{"ImportImageTasks": []}, {"ImportImageTasks": describe_import_image_tasks}],
+            [],
             {"changed": True, "msg": "Would have created the import task if not in check mode"},
         ),
         (
-            [{"ImportImageTasks": describe_import_image_tasks}, {"ImportImageTasks": describe_import_image_tasks}],
+            describe_import_image_tasks,
             {
                 "changed": False,
                 "msg": "An import task with the specified name already exists",
@@ -138,25 +122,28 @@ def test_present_no_check_mode(client, module, paginate, side_effects, expected_
         ),
     ],
 )
-def test_present_check_mode(client, module, paginate, side_effects, expected_result):
-    paginate.build_full_result.side_effect = side_effects
+@patch(module_name + ".import_image")
+@patch(module_name + ".describe_import_image_tasks_as_snake_dict")
+def test_present_check_mode(
+    m_describe_import_image_tasks_as_snake_dict, m_import_image, module, import_image_info, expected_result
+):
+    m_describe_import_image_tasks_as_snake_dict.return_value = import_image_info
     module.check_mode = True
     module.exit_json.side_effect = SystemExit(1)
+    connection = MagicMock()
 
-    with patch(utils + ".helper_describe_import_image_tasks", return_value=paginate):
-        with pytest.raises(SystemExit):
-            ec2_import_image.present(client, module)
+    with pytest.raises(SystemExit):
+        ec2_import_image.present(connection, module)
 
+    m_import_image.assert_not_called()
     module.exit_json.assert_called_with(**expected_result)
 
 
 @pytest.mark.parametrize(
-    "side_effect, expected_result",
+    "import_image_info, expected_result",
     [
         (
-            [
-                {"ImportImageTasks": []},
-            ],
+            [],
             {
                 "changed": False,
                 "msg": "The specified import task does not exist or it cannot be cancelled",
@@ -164,31 +151,38 @@ def test_present_check_mode(client, module, paginate, side_effects, expected_res
             },
         ),
         (
-            [
-                {"ImportImageTasks": describe_import_image_tasks},
-            ],
+            describe_import_image_tasks,
             {"changed": True, "import_image": expected_result},
         ),
     ],
 )
-def test_absent_no_check_mode(client, module, paginate, side_effect, expected_result):
-    paginate.build_full_result.side_effect = side_effect
+@patch(module_name + ".cancel_import_task")
+@patch(module_name + ".describe_import_image_tasks_as_snake_dict")
+def test_absent_no_check_mode(
+    m_describe_import_image_tasks_as_snake_dict, m_cancel_import_task, module, import_image_info, expected_result
+):
     module.exit_json.side_effect = SystemExit(1)
+    connection = MagicMock()
+    m_describe_import_image_tasks_as_snake_dict.return_value = import_image_info
+    m_cancel_import_task.return_value = True
 
-    with patch(utils + ".helper_describe_import_image_tasks", return_value=paginate):
-        with pytest.raises(SystemExit):
-            ec2_import_image.absent(client, module)
+    with pytest.raises(SystemExit):
+        ec2_import_image.absent(connection, module)
 
+    if import_image_info:
+        m_cancel_import_task.assert_called_with(
+            connection,
+            import_task_id=import_image_info[0]["import_task_id"],
+            cancel_reason=module.params.get("cancel_reason"),
+        )
     module.exit_json.assert_called_with(**expected_result)
 
 
 @pytest.mark.parametrize(
-    "side_effect, expected_result",
+    "import_image_info, expected_result",
     [
         (
-            [
-                {"ImportImageTasks": []},
-            ],
+            [],
             {
                 "changed": False,
                 "msg": "The specified import task does not exist or it cannot be cancelled",
@@ -196,21 +190,25 @@ def test_absent_no_check_mode(client, module, paginate, side_effect, expected_re
             },
         ),
         (
-            [
-                {"ImportImageTasks": describe_import_image_tasks},
-            ],
-            {"changed": True, "import_image": expected_result},
+            describe_import_image_tasks,
+            {"changed": True, "msg": "Would have cancelled the import task if not in check mode"},
         ),
     ],
 )
-def test_present_check_mode(client, module, paginate, side_effect, expected_result):
-    paginate.build_full_result.side_effect = side_effect
+@patch(module_name + ".cancel_import_task")
+@patch(module_name + ".describe_import_image_tasks_as_snake_dict")
+def test_absent_check_mode(
+    m_describe_import_image_tasks_as_snake_dict, m_cancel_import_task, module, import_image_info, expected_result
+):
     module.exit_json.side_effect = SystemExit(1)
+    module.check_mode = True
+    m_describe_import_image_tasks_as_snake_dict.return_value = import_image_info
+    connection = MagicMock()
 
-    with patch(utils + ".helper_describe_import_image_tasks", return_value=paginate):
-        with pytest.raises(SystemExit):
-            ec2_import_image.absent(client, module)
+    with pytest.raises(SystemExit):
+        ec2_import_image.absent(connection, module)
 
+    m_cancel_import_task.assert_not_called()
     module.exit_json.assert_called_with(**expected_result)
 
 
@@ -221,4 +219,4 @@ def test_main_success(m_AnsibleAWSModule):
 
     ec2_import_image_info.main()
 
-    m_module.client.assert_called_with("ec2", retry_decorator=ANY)
+    m_module.client.assert_called_with("ec2")
