+++ tests/unit/plugins/modules/test_ec2_snapshot_info.py
@@ -3,7 +3,6 @@
 # This file is part of Ansible
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from unittest.mock import ANY
 from unittest.mock import MagicMock
 from unittest.mock import call
 from unittest.mock import patch
@@ -30,61 +29,13 @@ def test_build_request_args(
     )
 
 
-def test_get_snapshots():
-    module = MagicMock()
-    connection = MagicMock()
-
-    connection.describe_snapshots.return_value = {
-        "Snapshots": [
-            {
-                "Description": "Created by CreateImage(i-083b9dd1234567890) for ami-01486e111234567890",
-                "Encrypted": False,
-                "OwnerId": "123456789000",
-                "Progress": "100%",
-                "SnapshotId": "snap-0f00cba1234567890",
-                "StartTime": "2021-09-30T01:04:49.724000+00:00",
-                "State": "completed",
-                "StorageTier": "standard",
-                "Tags": [
-                    {"Key": "TagKey", "Value": "TagValue"},
-                ],
-                "VolumeId": "vol-0ae6c5e1234567890",
-                "VolumeSize": 10,
-            },
-            {
-                "Description": "Created by CreateImage(i-083b9dd1234567890) for ami-01486e111234567890",
-                "Encrypted": False,
-                "OwnerId": "123456789000",
-                "Progress": "100%",
-                "SnapshotId": "snap-0f00cba1234567890",
-                "StartTime": "2021-09-30T01:04:49.724000+00:00",
-                "State": "completed",
-                "StorageTier": "standard",
-                "Tags": [
-                    {"Key": "TagKey", "Value": "TagValue"},
-                ],
-                "VolumeId": "vol-0ae6c5e1234567890",
-                "VolumeSize": 10,
-            },
-        ]
-    }
-
-    request_args = {"SnapshotIds": ["snap-0f00cba1234567890"]}
-
-    snapshot_info = ec2_snapshot_info.get_snapshots(connection, module, request_args)
-
-    assert connection.describe_snapshots.call_count == 1
-    connection.describe_snapshots.assert_called_with(aws_retry=True, SnapshotIds=["snap-0f00cba1234567890"])
-    assert len(snapshot_info["Snapshots"]) == 2
-
-
 @patch(module_name + ".build_request_args")
-@patch(module_name + ".get_snapshots")
-def test_list_ec2_snapshots(m_get_snapshots, m_build_request_args):
+@patch(module_name + ".describe_snapshots")
+def test_list_ec2_snapshots(m_describe_snapshots, m_build_request_args):
     module = MagicMock()
     connection = MagicMock()
 
-    m_get_snapshots.return_value = {
+    m_describe_snapshots.return_value = {
         "Snapshots": [
             {
                 "Description": "Created by CreateImage(i-083b9dd1234567890) for ami-01486e111234567890",
@@ -110,10 +61,10 @@ def test_list_ec2_snapshots(m_get_snapshots, m_build_request_args):
 
     ec2_snapshot_info.list_ec2_snapshots(connection, module, request_args)
 
-    assert m_get_snapshots.call_count == 1
-    m_get_snapshots.assert_has_calls(
+    assert m_describe_snapshots.call_count == 1
+    m_describe_snapshots.assert_has_calls(
         [
-            call(connection, module, m_build_request_args.return_value),
+            call(connection, **m_build_request_args.return_value),
         ]
     )
 
@@ -125,4 +76,4 @@ def test_main_success(m_AnsibleAWSModule):
 
     ec2_snapshot_info.main()
 
-    m_module.client.assert_called_with("ec2", retry_decorator=ANY)
+    m_module.client.assert_called_with("ec2")
